Optimize frame grabber + Embeded CC suport

+++ FFmpegInterop/Source/FFmpegReader.cpp
@@ -49,7 +49,7 @@ int FFmpegReader::ReadPacket()
 		return ret;
 	}
 
-	if (avPacket->stream_index > sampleProviders->size() || avPacket->stream_index < 0)
+	if (avPacket->stream_index > (int)sampleProviders->size() || avPacket->stream_index < 0)
 	{
 		av_packet_free(&avPacket);
 		return E_FAIL;
